fix: expose torch-like functions, add more tests
