+++ src/shift_nth_row_n_steps/__init__.py
@@ -1,4 +1,5 @@
 __version__ = "0.1.2"
 from ._main import shift_nth_row_n_steps
+from ._torch_like import narrow, select, take_slice
 
-__all__ = ["shift_nth_row_n_steps"]
+__all__ = ["shift_nth_row_n_steps", "narrow", "select", "take_slice"]
