--- src/shift_nth_row_n_steps/main.py
@@ -1,3 +1,0 @@
-def add(n1: int, n2: int) -> int:
-    """Add the arguments."""
-    return n1 + n2
