+++ tests/test_torch_like.py
@@ -1,0 +1,23 @@
+import importlib.util
+from unittest import SkipTest
+
+import pytest
+import torch
+
+from shift_nth_row_n_steps import narrow, select
+
+
+@pytest.fixture
+def setup() -> None:
+    if importlib.util.find_spec("torch") is None:
+        raise SkipTest("torch is not installed")
+
+
+def test_select() -> None:
+    input = torch.randn(3, 4, 5)
+    assert torch.equal(select(input, 0, axis=0), torch.select(input, 0, 0))
+
+
+def test_narrow() -> None:
+    input = torch.randn(3, 4, 5)
+    assert torch.equal(narrow(input, 1, 2, axis=0), torch.narrow(input, 0, 1, 2))
